型が増えそうなので分割。enemyAction関連の増加が見込まれる
