+++ src/rmmz/rpg/data/main/traitContainers/enemy/types/action.ts
@@ -1,0 +1,7 @@
+export interface Enemy_Action {
+  conditionParam1: number;
+  conditionParam2: number;
+  conditionType: number;
+  skillId: number;
+  rating: number;
+}
