+++ src/rmmz/rpg/data/main/traitContainers/enemy/types/index.ts
@@ -1,0 +1,2 @@
+export * from "./action";
+export * from "./types";
