+++ src/rmmz/rpg/data/main/traitContainers/enemy/types/types.ts
@@ -1,5 +1,6 @@
-import type { ParamArray } from "../members";
-import type { Data_TraitContainer, Trait } from "../trait";
+import type { ParamArray } from "../../members";
+import type { Data_TraitContainer, Trait } from "../../trait";
+import type { Enemy_Action } from "./action";
 
 export interface Data_Enemy extends Data_TraitContainer {
   name: string;
@@ -16,13 +17,6 @@ export interface Data_Enemy extends Data_TraitContainer {
   dropItems: DropItem[];
 }
 
-export interface Enemy_Action {
-  conditionParam1: number;
-  conditionParam2: number;
-  conditionType: number;
-  skillId: number;
-  rating: number;
-}
 export interface DropItem {
   dataId: number;
   denominator: number;
